nusuk-converter/backend/requirements.txt
amqp==5.3.1
billiard==4.2.2
blinker==1.9.0
celery==5.4.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
gevent==25.9.1
gpt_context_builder==1.0.2
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.2
packaging==25.0
pathspec==0.12.1
pillow==10.4.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.4
regex==2025.9.1
requests==2.32.5
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.44
stripe==10.4.0
tiktoken==0.11.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
zope.event==6.0
zope.interface==8.0.1
gunicorn==22.0.0